backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.27.0
pydantic>=2.11.7
python-multipart>=0.0.6

# RAG (기존 — 레거시 호환용)
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
tiktoken>=0.8.0

# Agent LLM
openai>=1.0.0

# 문서 파서
pypdf>=5.0.0
python-docx>=1.1.0
ddgs
